Hannes' solution (for now)

+++ source/BeerSender/BeerSender.Domain/Boxes/Bottle.cs
@@ -3,5 +3,5 @@
 public record Bottle(
     string Brewery,
     string Name,
-    int Alcohol_percentage,
+    decimal Alcohol_percentage,
     int Volume_in_ml);
